+++ MonacoEditorComponent/Monaco/Helpers/CssInlineStyle.cs
@@ -4,18 +4,8 @@ using Windows.UI.Xaml.Media;
 
 namespace Monaco.Helpers
 {
-    public enum TextDecoration
-    {
-        None,
-        Underline,
-        Overline,
-        LineThrough, // line-through
-        Initial,
-        Inherit
-    }
-
     // Inline styles modify the text style itself and are useful for manipulating the colors and styles of text to indicate conditions.
-    public sealed class CssInlineStyle: ICssStyle
+    public sealed class CssInlineStyle : ICssStyle
     {
         public TextDecoration TextDecoration { get; set; }
         public FontWeight? FontWeight { get; set; }
@@ -29,9 +19,9 @@ namespace Monaco.Helpers
 
         public string Name { get; private set; }
 
-        public CssInlineStyle()
+        public CssInlineStyle(CodeEditor editor)
         {
-            Name = CssStyleBroker.Instance.Register(this);
+            Name = CssStyleBroker.GetInstance(editor).Register(this);
         }
 
         public string ToCss()